+++ databruce/config/prod.py
@@ -86,6 +86,7 @@ LOGGING = {
 sentry_sdk.init(
     dsn=os.getenv("SENTRY_DSN"),
     send_default_pii=True,
+    traces_sample_rate=1.0,
 )
 
 # Database